hotfix hashing

+++ src/hotfixes/processing.h
@@ -6,7 +6,7 @@ namespace dhf::hotfixes {
 struct FJsonObject;
 
 extern const std::string& running_hotfix_name;
-extern const size_t& running_hotfix_hash;
+extern const uint64_t& running_hotfix_hash;
 
 /**
  * @brief Handles `GbxSparkSdk::Discovery::Services::FromJson` calls, inserting our custom hotfixes.
